New Name - CardDefinitions.js
New FileName, this js file will hold only some base descriptions of cards, items, and the deck. 
Cards are generally for battle, items are generally for consuming or story interaction, to be included later when I can make a story.
The deck itself is an object that holds an array of card objects, where the array obviously can hold different kinds of cards.