Item Descriptions
A library of Item Descriptions, should be able to define all objects and cards that the player can interact with
Environment objects, most likely will be given an eventlistener of click to give a message of what kind of "observation" or "interaction" the player makes with the object (i.e. click on a wagon of apples. "It's a wagon full of delicious Cebelrai apples".
Battle Cards, which can be taken as a reward, may be able to find within the environment as well.
Consumables, key consumables are for story progression or quests. battle consumables