Player.js
Here is where the Player objects are stored,
A base entity to state that this is a person and not an object, though some entities will be treated as part of the environment, as in interactable but not collectible or battle.
enemies base class, stats are a part of this class so we can battle and compare numbers. 
player class, a player will accept default name or give custom name. preset stats while testing.  
then we start to define enemies, such as whinter snek.